requirements.txt
pandas
pylint
pytest
pytest-cov
bs4
requests
python-dotenv